refactor power control tab to use StateFlow exposed from PowerManager
