+++ app/src/main/java/app/gamenative/ui/component/quickMenus/PowerControlQuickMenuTab.kt
@@ -4,73 +4,40 @@ import androidx.compose.foundation.focusGroup
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.State
-import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.tooling.preview.Preview
-import app.gamenative.powercontrol.AutoTuningStrategy
+import app.gamenative.powercontrol.CpuDisplayInfo
+import app.gamenative.powercontrol.GpuDisplayInfo
+import app.gamenative.powercontrol.PowerControlUiState
 import app.gamenative.powercontrol.PowerManager
 import app.gamenative.powercontrol.PowerProfile
+import app.gamenative.powercontrol.RamDisplayInfo
 import app.gamenative.powercontrol.profiles.CpuGovernor
 import app.gamenative.powercontrol.profiles.PerformancePreset
-import app.gamenative.powercontrol.PowerProfiles
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import timber.log.Timber
-
-sealed class PowerControlUiState {
-    object Loading : PowerControlUiState()
-    data class Success(
-        val cpuInfo: CpuDisplayInfo,
-        val gpuInfo: GpuDisplayInfo?,
-        val ramInfo: RamDisplayInfo?,
-        val selectedProfile: PowerProfile,
-        val availableProfiles: List<PowerProfile>
-    ) : PowerControlUiState()
-}
-
-data class CpuDisplayInfo(
-    val currentGovernor: String,
-    val availableGovernors: List<String>,
-    val availableFrequencies: List<Long>,
-    val currentMinValue: Long,
-    val currentMaxValue: Long,
-    val selectedMinFreqIndex: Int,
-    val selectedMaxFreqIndex: Int
-)
-
-data class GpuDisplayInfo(
-    val availableFrequencies: List<Long>,
-    val currentFreqIndex: Int,
-    val minPowerLevel: Int,
-    val maxPowerLevel: Int,
-    val maxAvailablePowerLevel: Int
-)
-
-data class RamDisplayInfo(
-    val minBusLevel: Int,
-    val maxBusLevel: Int,
-    val maxAvailableBusLevel: Int
-)
 
 @Composable
 fun PowerControlQuickMenuTab(
     modifier: Modifier = Modifier,
     focusRequester: FocusRequester? = null,
 ) {
-    var refreshTrigger by remember { mutableIntStateOf(0) }
-    val uiState by rememberPowerControlState(refreshTrigger)
+    val uiState by PowerManager.uiState.collectAsState()
     val coroutineScope = rememberCoroutineScope()
     val isDriverSupported = remember { PowerManager.isDriverSupported() }
+
+    LaunchedEffect(Unit) {
+        withContext(Dispatchers.IO) {
+            PowerManager.refreshUiState()
+        }
+    }
 
     PowerControlQuickMenuContent(
         uiState = uiState,
@@ -80,7 +47,7 @@ fun PowerControlQuickMenuTab(
                 PowerManager.currentProfile?.let { profile ->
                     PowerManager.setCurrentProfile(profile.copy(enableFanControl = enabled))
                 }
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onGamePinningToggled = { enabled ->
@@ -88,7 +55,7 @@ fun PowerControlQuickMenuTab(
                 PowerManager.currentProfile?.let { profile ->
                     PowerManager.setCurrentProfile(profile.copy(enableGamePinning = enabled))
                 }
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onTuningModeSelected = { perCluster ->
@@ -96,7 +63,7 @@ fun PowerControlQuickMenuTab(
                 PowerManager.currentProfile?.let { profile ->
                     PowerManager.setCurrentProfile(profile.copy(enablePerClusterTuning = perCluster))
                 }
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onAdaptiveFpsCapToggled = { enabled ->
@@ -104,7 +71,7 @@ fun PowerControlQuickMenuTab(
                 PowerManager.currentProfile?.let { profile ->
                     PowerManager.setCurrentProfile(profile.copy(enableAdaptiveFpsCap = enabled))
                 }
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onAutoTuningToggled = { enabled ->
@@ -118,7 +85,7 @@ fun PowerControlQuickMenuTab(
                     PowerManager.setCurrentProfile(updatedProfile)
                 }
 
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onTuningStrategySelected = { strategy ->
@@ -132,7 +99,7 @@ fun PowerControlQuickMenuTab(
                     PowerManager.setCurrentProfile(updatedProfile)
                 }
 
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onProfileSelected = { profile ->
@@ -167,14 +134,14 @@ fun PowerControlQuickMenuTab(
                 }
 
                 Timber.d("Profile application result: $success")
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onGovernorSelected = { governor ->
             coroutineScope.launch(Dispatchers.IO) {
                 PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                 PowerManager.setGovernor(governor)
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onMinCpuValueChanged = { freqIndex ->
@@ -183,7 +150,7 @@ fun PowerControlQuickMenuTab(
                 coroutineScope.launch(Dispatchers.IO) {
                     PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                     PowerManager.setMinCpuValue(freq)
-                    refreshTrigger++
+                    PowerManager.refreshUiState()
                 }
             }
         },
@@ -193,7 +160,7 @@ fun PowerControlQuickMenuTab(
                 coroutineScope.launch(Dispatchers.IO) {
                     PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                     PowerManager.setMaxCpuValue(freq)
-                    refreshTrigger++
+                    PowerManager.refreshUiState()
                 }
             }
         },
@@ -201,176 +168,33 @@ fun PowerControlQuickMenuTab(
             coroutineScope.launch(Dispatchers.IO) {
                 PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                 PowerManager.setMinGpuPowerLevel(powerLevel)
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onMaxGpuPowerChanged = { powerLevel ->
             coroutineScope.launch(Dispatchers.IO) {
                 PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                 PowerManager.setMaxGpuPowerLevel(powerLevel)
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onMinRamValueChanged = { powerLevel ->
             coroutineScope.launch(Dispatchers.IO) {
                 PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                 PowerManager.setMinBusLevel(powerLevel)
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         onMaxRamValueChanged = { powerLevel ->
             coroutineScope.launch(Dispatchers.IO) {
                 PowerManager.setProfileName(PerformancePreset.CUSTOM.displayName)
                 PowerManager.setMaxBusLevel(powerLevel)
-                refreshTrigger++
+                PowerManager.refreshUiState()
             }
         },
         firstItemFocusRequester = focusRequester,
         modifier = modifier.focusGroup(),
     )
-}
-
-@Composable
-private fun rememberPowerControlState(refreshTrigger: Int): State<PowerControlUiState> {
-    var uiState by remember { mutableStateOf<PowerControlUiState>(PowerControlUiState.Loading) }
-    var selectedProfile by remember {
-        mutableStateOf(
-            PowerProfile(
-                name = PerformancePreset.CUSTOM.displayName,
-                governor = CpuGovernor.SCHEDUTIL,
-                minCpuFreq = 0,
-                maxCpuFreq = 0,
-                minGpuPowerLevel = 0,
-                maxGpuPowerLevel = 0,
-                minBusLevel = 0,
-                maxBusLevel = 0,
-            )
-        )
-    }
-    var isInitialized by remember { mutableStateOf(false) }
-
-    LaunchedEffect(refreshTrigger) {
-        withContext(Dispatchers.IO) {
-            try {
-                val hasGpuSupport = PowerManager.isGpuSupported()
-                val hasBusSupport = PowerManager.isBusSupported()
-
-                val cpuInfo = PowerManager.getCpuInfo()
-                if (cpuInfo != null) {
-                    val availableGovernors = PowerManager.getAvailableGovernors()
-                    val availableFrequencies = PowerManager.getAvailableCpuFrequencies()
-
-                    val gpuDisplayInfo = if (hasGpuSupport) {
-                        val gpuInfo = PowerManager.getGpuInfo()
-                        val availableGpuFrequencies = PowerManager.getAvailableGpuFrequencies()
-                        if (gpuInfo != null) {
-                            val maxGpuPowerLevel = if (gpuInfo.numGpuPowerLevels > 0) {
-                                gpuInfo.numGpuPowerLevels - 1
-                            } else 0
-                            val currentFreqIndex = if (availableGpuFrequencies.isNotEmpty()) {
-                                availableGpuFrequencies.indexOfFirst {
-                                    it >= gpuInfo.currentGpuValue
-                                }.coerceAtLeast(0)
-                            } else {
-                                0
-                            }
-                            GpuDisplayInfo(
-                                availableFrequencies = availableGpuFrequencies,
-                                currentFreqIndex = currentFreqIndex,
-                                minPowerLevel = gpuInfo.minGpuPowerLevel,
-                                maxPowerLevel = gpuInfo.maxGpuPowerLevel,
-                                maxAvailablePowerLevel = maxGpuPowerLevel
-                            )
-                        } else null
-                    } else null
-
-                    val ramDisplayInfo = if (hasBusSupport) {
-                        val busInfo = PowerManager.getBusInfo()
-
-                        if (busInfo != null && busInfo.numBusLevels > 0) {
-                            RamDisplayInfo(
-                                minBusLevel = busInfo.minBusLevel,
-                                maxBusLevel = busInfo.maxBusLevel,
-                                maxAvailableBusLevel = busInfo.numBusLevels - 1
-                            )
-                        } else {
-                            null
-                        }
-                    } else {
-                        null
-                    }
-
-                    val selectedMinFreqIndex = availableFrequencies.indexOfFirst {
-                        it >= cpuInfo.currentMinValue
-                    }.coerceAtLeast(0)
-                    val selectedMaxFreqIndex = availableFrequencies.indexOfFirst {
-                        it >= cpuInfo.currentMaxValue
-                    }.coerceAtLeast(0)
-
-                    val maxGpuPowerLevel = gpuDisplayInfo?.maxAvailablePowerLevel ?: 0
-                    val profiles = PowerProfiles.getDefaultProfiles(availableGovernors, availableFrequencies, maxGpuPowerLevel)
-
-                    val currentGovernor = CpuGovernor.fromString(cpuInfo.currentGovernor)
-
-                    Timber.d("Current profile: $selectedProfile")
-                    profiles.forEach { profile ->
-                        Timber.d("Profile $profile")
-                    }
-
-                    // Try to match current settings against available profiles
-                    // Match by PowerManager's current profile name
-                    val matchingProfile = profiles.find { profile ->
-                        profile.name == (PowerManager.currentProfile?.name ?: PerformancePreset.CUSTOM.displayName)
-                    }
-
-                    Timber.d("Matching profile: $matchingProfile")
-
-                    selectedProfile = (matchingProfile ?: PowerProfile(
-                        name = PerformancePreset.CUSTOM.displayName,
-                        governor = currentGovernor ?: CpuGovernor.SCHEDUTIL,
-                        minCpuFreq = cpuInfo.currentMinValue,
-                        maxCpuFreq = cpuInfo.currentMaxValue,
-                        minGpuPowerLevel = gpuDisplayInfo?.minPowerLevel ?: 0,
-                        maxGpuPowerLevel = gpuDisplayInfo?.maxPowerLevel ?: 0,
-                        minBusLevel = ramDisplayInfo?.minBusLevel ?: 0,
-                        maxBusLevel = ramDisplayInfo?.maxBusLevel ?: 0
-                    )).copy(
-                        // Preserve enableAutoTuning and tuningStrategy from PowerManager's current profile
-                        enableAutoTuning = PowerManager.currentProfile?.enableAutoTuning ?: true,
-                        enablePerClusterTuning = PowerManager.currentProfile?.enablePerClusterTuning ?: true,
-                        enableAdaptiveFpsCap = PowerManager.currentProfile?.enableAdaptiveFpsCap ?: true,
-                        enableFanControl = PowerManager.currentProfile?.enableFanControl ?: true,
-                        enableGamePinning = PowerManager.currentProfile?.enableGamePinning ?: false,
-                        tuningStrategy = PowerManager.currentProfile?.tuningStrategy ?: AutoTuningStrategy.POWER_EFFICIENT
-                    )
-
-                    if (!isInitialized) {
-                        isInitialized = true
-                    }
-
-                    uiState = PowerControlUiState.Success(
-                        cpuInfo = CpuDisplayInfo(
-                            currentGovernor = cpuInfo.currentGovernor,
-                            availableGovernors = availableGovernors,
-                            availableFrequencies = availableFrequencies,
-                            currentMinValue = cpuInfo.currentMinValue,
-                            currentMaxValue = cpuInfo.currentMaxValue,
-                            selectedMinFreqIndex = selectedMinFreqIndex,
-                            selectedMaxFreqIndex = selectedMaxFreqIndex
-                        ),
-                        gpuInfo = gpuDisplayInfo,
-                        selectedProfile = selectedProfile,
-                        availableProfiles = profiles,
-                        ramInfo = ramDisplayInfo,
-                    )
-                }
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
-        }
-    }
-
-    return remember { derivedStateOf { uiState } }
 }
 
 @Preview(showBackground = true, name = "Loading State")
